refactor(capabilities): group framework into core/, keep verbs top-level
