+++ surfsense_backend/app/capabilities/core/__init__.py
@@ -1,0 +1,30 @@
+"""Capability framework kernel: registry contracts, store, billing, and access doors."""
+
+from app.capabilities.core.billing import charge_capability, gate_capability
+from app.capabilities.core.store import (
+    all_capabilities,
+    get_capability,
+    register_capability,
+)
+from app.capabilities.core.types import (
+    BillableInput,
+    BillableOutput,
+    BillingUnit,
+    Capability,
+    CapabilityContext,
+    Executor,
+)
+
+__all__ = [
+    "BillableInput",
+    "BillableOutput",
+    "BillingUnit",
+    "Capability",
+    "CapabilityContext",
+    "Executor",
+    "all_capabilities",
+    "charge_capability",
+    "gate_capability",
+    "get_capability",
+    "register_capability",
+]
