+++ surfsense_backend/app/routes/__init__.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends
 
 import app.capabilities.web  # noqa: F401  (registers web.* verbs before the door builds)
 from app.automations.api import router as automations_router
-from app.capabilities.access.rest import build_capabilities_router
+from app.capabilities.core.access.rest import build_capabilities_router
 from app.file_storage.api import router as file_storage_router
 from app.gateway import require_gateway_enabled
 from app.notifications.api import router as notifications_router
